разделил основной файл main.js на модули
