+++ js/db.js
@@ -1,0 +1,34 @@
+const AVATAR_COUNT = 8;
+const TYPE_HOUSE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+const CHECKINS = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const CHECKOUTS = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+const COUNT_OFFER = 10;
+
+export {AVATAR_COUNT, TYPE_HOUSE, CHECKINS, CHECKOUTS, FEATURES, PHOTOS, COUNT_OFFER};
